Rename SOS-related files and directories to use `Sos` instead of `SOS`
