+++ src/core/Decode/DecodeSos/DecodeSos.c
@@ -1,4 +1,4 @@
-#include "DecodeSOS.h"
+#include "DecodeSos.h"
 #include "../UtilDecode/UtilDecode.h"
 #include "../../UtilCore/UtilCore.h"
 #include "../../../data/md1database/md1database.h"
