+++ src/core/Encode/EncodeSos/EncodeSos.c
@@ -1,4 +1,4 @@
-#include "EncodeSOS.h"
+#include "EncodeSos.h"
 #include "../UtilEncode/UtilEncode.h"
 #include "../../UtilCore/UtilCore.h"
 #include "../../../data/md1database/md1database.h"
