+++ src/core/Encode/EncodeSos/EncodeSos.h
